Fix sqlite session

+++ telethon_secret_chat/storage/sqlite.py
@@ -39,7 +39,7 @@ class SecretSQLiteSession(SecretMemorySession):
             self._create_table(
                 c,
                 f"""{TABLE_NAME} (
-                  id integer,
+                  id integer NOT NULL PRIMARY KEY,
                   access_hash integer,
                   auth_key blob,
                   admin integer,
@@ -93,7 +93,6 @@ class SecretSQLiteSession(SecretMemorySession):
 
     def save_chat(self, chat: SecretChat, temp=False):
         c = self._conn.cursor()
-
         row = (
             chat.id, chat.access_hash, chat.auth_key, 1 if chat.admin else 0, chat.user_id, chat.in_seq_no_x,
             chat.out_seq_no_x, chat.in_seq_no, chat.out_seq_no, chat.layer, chat.ttl, chat.ttr, chat.updated,
@@ -107,29 +106,29 @@ class SecretSQLiteSession(SecretMemorySession):
 
     def get_temp_secret_chat_by_id(self, id):
         row = self._execute(
-            f"select * from {TABLE_NAME} where temp=1 and id = ?", id)
+            f"select * from {TABLE_NAME} where temp=1 and id = ? or user_id=?", id, id)
         if row:
             input_chat = InputEncryptedChat(chat_id=row[0], access_hash=row[1])
             return SecretChat(id=row[0], access_hash=row[1], auth_key=row[2], admin=True if row[3] else False,
                               user_id=row[4], in_seq_no_x=row[5], out_seq_no_x=row[6], in_seq_no=row[7],
                               out_seq_no=row[8], layer=row[9], ttl=row[10], ttr=row[11], updated=row[12],
-                              created=row[13], mtproto=row[14], input_chat=input_chat)
+                              created=row[13], mtproto=row[14], input_chat=input_chat, session=self, is_temp=True)
 
     def get_secret_chat_by_id(self, id):
         row = self._execute(
-            f"select * from {TABLE_NAME} where temp=0 and id = ?", id)
+            f"select * from {TABLE_NAME} where temp=0 and id = ? or user_id = ? ORDER BY UPDATED", id, id)
 
         if row:
             input_chat = InputEncryptedChat(chat_id=row[0], access_hash=row[1])
             return SecretChat(id=row[0], access_hash=row[1], auth_key=row[2], admin=True if row[3] else False,
                               user_id=row[4], in_seq_no_x=row[5], out_seq_no_x=row[6], in_seq_no=row[7],
                               out_seq_no=row[8], layer=row[9], ttl=row[10], ttr=row[11], updated=row[12],
-                              created=row[13], mtproto=row[14], input_chat=input_chat)
+                              created=row[13], mtproto=row[14], input_chat=input_chat, session=self)
 
     def remove_secret_chat_by_id(self, id, temp=False):
         print("removing chat with id", id, "and is ", temp)
         c = self._conn.cursor()
         try:
-            c.execute(f"delete from {TABLE_NAME} where id=? and temp=?", (id, 1 if temp else 0))
+            c.execute(f"delete from {TABLE_NAME} where id=? or user_id=? and temp=?", (id, id, 1 if temp else 0))
         finally:
             c.close()
